search_img/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
PyYAML>=6.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.3.0
Pillow>=10.0.0
